organise validation code
