+++ src/app.ts
@@ -1,54 +1,8 @@
 /// <reference path = "drag-drop-interfaces.ts"/>
 /// <reference path = "postit-model.ts"/>
+/// <reference path = "postit-state.ts"/>
 
 namespace App {
-  //decorators
-  interface Validatable {
-    value: string | number;
-    required?: boolean;
-    minLength?: number;
-    maxLength?: number;
-
-    minValue?: number;
-    maxValue?: number;
-  }
-
-  // Project state Management
-  type Listener = (items: Postit[]) => void;
-
-  function validate(validatableInput: Validatable): boolean {
-    let isValid = true;
-
-    if (validatableInput.required) {
-      isValid =
-        isValid && validatableInput.value.toString().trim().length !== 0;
-    }
-
-    if (validatableInput.minLength != null) {
-      isValid =
-        isValid &&
-        validatableInput.value.toString().trim().length >=
-          validatableInput.minLength;
-    }
-
-    if (validatableInput.maxLength != null) {
-      isValid =
-        isValid &&
-        validatableInput.value.toString().trim().length <=
-          validatableInput.maxLength;
-    }
-
-    if (validatableInput.minValue != null) {
-      isValid = isValid && validatableInput.value >= validatableInput.minValue;
-    }
-
-    if (validatableInput.maxValue != null) {
-      isValid = isValid && validatableInput.value <= validatableInput.maxValue;
-    }
-
-    return isValid;
-  }
-
   function autobind(_: any, _2: string, descriptor: PropertyDescriptor) {
     const originalMethod = descriptor.value;
     const adjDescriptor: PropertyDescriptor = {
@@ -252,94 +206,6 @@ namespace App {
       this.element.querySelector("p")!.textContent = this.postit.description;
     }
   }
-
-  class Postits {
-    private listeners: Listener[] = [];
-
-    postitList: Postit[] = [];
-
-    hostElement?: HTMLElement;
-    postitTemplate?: HTMLTemplateElement;
-    postitElement?: HTMLElement;
-
-    private static _Postits: Postits;
-
-    private constructor() {}
-
-    static getInstance() {
-      if (this._Postits) {
-        return this._Postits;
-      }
-      this._Postits = new Postits();
-
-      return this._Postits;
-    }
-
-    addListener(listenerFn: Listener) {
-      this.listeners.push(listenerFn);
-    }
-
-    addPostit(postit: Postit): boolean {
-      const validation = this.Validation(postit);
-
-      if (validation) {
-        this.postitList.push(postit);
-
-        this.updateListeners();
-      }
-
-      return validation;
-    }
-
-    movePostit(id: string, newStatus: Status) {
-      const postit = this.postitList.filter((postit) => postit.id === id)[0];
-
-      if (postit && postit.status !== newStatus) {
-        postit.status = newStatus;
-        this.updateListeners();
-      }
-    }
-
-    private updateListeners() {
-      for (const listenerFn of this.listeners) {
-        listenerFn(this.postitList.slice());
-      }
-    }
-
-    private Validation(project: Postit): boolean {
-      const titleValidatable: Validatable = {
-        value: project.title,
-        required: true,
-        minLength: 1,
-        maxLength: 50,
-      };
-
-      const descriptionValidatable: Validatable = {
-        value: project.description,
-        required: true,
-        minLength: 1,
-        maxLength: 200,
-      };
-
-      const valueValidatable: Validatable = {
-        value: project.status,
-        required: true,
-        minValue: 0,
-        maxValue: 3,
-      };
-
-      if (
-        !validate(titleValidatable) ||
-        !validate(descriptionValidatable) ||
-        !validate(valueValidatable)
-      ) {
-        return false;
-      }
-      return true;
-    }
-  }
-
-  const postits = Postits.getInstance();
 
   new InputComponent();
   new PostitListComponent(Status.Unassigned);
