+++ src/validation.ts
@@ -1,0 +1,46 @@
+namespace App {
+  export interface Validatable {
+    value: string | number;
+    required?: boolean;
+    minLength?: number;
+    maxLength?: number;
+
+    minValue?: number;
+    maxValue?: number;
+  }
+
+  // Project state Management
+
+  export function validate(validatableInput: Validatable): boolean {
+    let isValid = true;
+
+    if (validatableInput.required) {
+      isValid =
+        isValid && validatableInput.value.toString().trim().length !== 0;
+    }
+
+    if (validatableInput.minLength != null) {
+      isValid =
+        isValid &&
+        validatableInput.value.toString().trim().length >=
+          validatableInput.minLength;
+    }
+
+    if (validatableInput.maxLength != null) {
+      isValid =
+        isValid &&
+        validatableInput.value.toString().trim().length <=
+          validatableInput.maxLength;
+    }
+
+    if (validatableInput.minValue != null) {
+      isValid = isValid && validatableInput.value >= validatableInput.minValue;
+    }
+
+    if (validatableInput.maxValue != null) {
+      isValid = isValid && validatableInput.value <= validatableInput.maxValue;
+    }
+
+    return isValid;
+  }
+}
